+++ backend/src/models/hypixel/inventory.rs
@@ -2,12 +2,6 @@ use std::collections::HashMap;
 
 use serde::Deserialize;
 
-use crate::{
-    models::{self},
-    processing,
-    routes::ApiError,
-};
-
 #[derive(Deserialize)]
 pub struct Inventory {
     #[serde(rename = "type")]
@@ -98,12 +92,4 @@ pub struct ItemSkullOwnerProperty {
 pub struct ItemSkullOwnerTexture {
     #[serde(rename = "Value")]
     pub value: String,
-}
-
-pub fn process_optional_inventory(
-    inv: &Option<models::hypixel::inventory::Inventory>,
-) -> Result<Option<Vec<Option<models::profile::Item>>>, ApiError> {
-    inv.as_ref()
-        .map(processing::inventory::inventory)
-        .transpose()
 }
