+++ backend/src/models/mod.rs
@@ -1,4 +1,8 @@
+//! This module contains all the structs used by our API and the Hypixel API. It does not contain
+//! any functions, those go in [`crate::processing`].
+
 pub mod error;
 pub mod hypixel;
+pub mod inventory;
 pub mod player;
 pub mod profile;
